coord: Reduce read latency (#12958)
Every one second the global timeline is advanced to the current time,
if the current time is ahead of the global timeline. This puts the
global timeline in write mode, which means that the next read will have
to advance the table when it's complete. This is undesirable because
some reads need to wait for two table advancements.

This commit explicitly puts the global timeline in read mode after
periodic advancements, which prevents reads that happen after the
advancement from having to advance the table again.

Writes will always have to advance a table regardless of what mode
the global timeline is in, so there are no negative impacts on writes.

Works towards resolving #12944