coord: split up advance table processing
Previously, all tables were advanced in the same append call. The
implementation at the bottom of that was serial, so coord would be
blocked until each table was processed.

Break this up so that each table is processed one at a time at the lowest
message priority. This means that incoming user queries are blocked for
at most the time it takes to process a single table advance instead of
all of them.

As a way to exploit parallelism or batched performance, use an
adaptive batch size when submitting advancement requests with a target
duration. This should automate a balance between user responsiveness
and total processing time.

Use a data structure that allows for the 1s advance loop to trigger faster
than it takes us to advance all of the tables. This could happen if:

- a user has many tables
- there are many mz instances all using the same postgres instance and its load increases query times
- the postgres instance hangs for a few seconds
- network or other delays somewhere in the stack

This solution is much simpler than attempting to split off the storage
controller into its own task.