dataflow-types,sql: don't hide connector enum behind trait
This is the next work item in source connectors (#11504). The
`Connector` enum isn't something that can be meaningfully abstracted,
since each connector type has such a distinct set of fields. So just
expose that enum directly to the SQL layer, rather than trying to hide
it behind a trait. This will fix at least one of the bugs in #12472.

Fix #12982.