materialized: rename to environmentd
`materialized` is now but one component of Materialize Cloud. Rename it
to `environmentd` to reflect its new purpose of managing a single
environment. The name `materialized` lives on as the name of the
all-in-one Docker image that can be run locally and in CI.

This commit also removes the Debian and tarball packaging of
`materialized`; `materialized`/`environmentd` haven't been runnable on
their own in a while.

Fix #3188.