sql: fix handling of SSL secrets
Secrets store actual secret values, not the path to a file on disk.
Update the Kafka and Confluent Schema Registry connections to use the
options that accept SSL certificates and keys as strings rather than
file paths; e.g., where we used to provide/set the `ssl.key.location`
option, we now provide/set the `ssl.key.pem` option.

Kafka/CSR options are now classified into three types:

  * Options which must *not* be a secret. These are options like
    fetch_message_max_bytes where we need to perform validation on the
    value at plan time.
  * Options which *must* be a secret. These are options like SSL keys
    where we want to force users to do the right thing and securely
    store their credentials.
  * Options which may or may not be a secret. These are options like
    usernames and SSL certificates which are generally not sensitive,
    but are allowed to be secrets so that they can be managed uniformly
    with SSL keys and passwords.

The bulk of this commit is updating the test files for the change.

This is another step towards #11504.