docs: Separate correctness prop 1 into two props
Originally we wanted strict serializability for acknowledgements of
upstream sources. However, we decided that this property was too strong
for our uses, due to its negative availability ramifications. Once an
acknowledgement has been sent by Materialize, then no reads can be
serviced until the upstream source has confirmed that it has received
our acknowledgement. Before the confirmation, Materialize has no way of
knowing if the acknowledgement has been received, and any timestamp
selected for a read will be susceptible to some consistency violation.

For now we have removed this correctness property and changed it so
that upstream acknowledgements only indicate that the data has been
made durable. In the future we may want to include some consistency
guarantees around upstream acknowledgements. Specifically it could be
useful for synchronous replication with upstream sources.

Fixes #13018