persist: add a flag to open-loop benchmark to write metrics out
A new registry is created at the beginning of the run and this is
written out after all the workers have been joined, so it exactly
contains our production prometheus metrics for one run of the benchmark.
This is immediately useful for compaction because it will let me get a
sense for real-world-ish write amplification.

There's all sorts of fancy things we could build on top of this, like
diffing vs previous runs, but for now just start with writing them out
so I can eyeball them.