secrets: add SecretsReader to ConnectionContext
This is a small step towards #13017. By putting the `SecretsReader` in
the `ConnectionContext`, we'll have easy access to it in `storaged` (and
`computed`, while sinks still live there) at the moment that we go to
instantiate a Kafka or Confluent Schema Registry connection. The
orchestrators are wired up to pass a new `--secrets-path` argument to
`storaged` and `computed` that tells them where secrets are mounted.