sql: better plan for single table function in scalar position
Table functions (set-returning functions) in scalar position are
currently planned as if they were inside a `ROWS FROM` lateral join,
as Postgres does. The issue fixed in this work arises when there is
a single function, which is indistinguishable from a regular lateral
join, but still used to be planned as if using `ROWS FROM`, which
involves adding ordinality information via `row_number()`, which
unfortunately was quite expensive computationally.

More context in #10801.