persist: pass around Metrics in an Arc, but not of the FooMetrics
I almost did this on the original PR, but didn't notice how clunky these
were until after aljoscha gave his final sign-off before vacation and I
didn't want to mess with it too much between then and merging. This
situation gets much worse in my local compaction branch, so getting
ahead of this now.

Specifically, the rule I used here was that it was fine to keep passing
a &FooMetrics as a ref, but if it needed to be an Arc, I passed the
top-level Arc<Metrics> instead.

While I'm in here, also make these pub instead of pub(crate), because it
will be useful to programmatically query them in persistcli tools.