mzcompose: Allow multiple Mz containers on the same host
For RQG comparison testing, two Mz versions need to be running
in a manner that also exposes them both on the host. For the purpose,
Materialized() instances need to be able to pass a 'HOSTPORT:CONTAINERPORT'
pair as part of their service definitions while keeping all the
internal Mz and Pg ports private.