coord: Fix group commit write locking
Previously, a write would try and lock the global write lock before
submitting the write for group commit. If it was unable to acquire the
lock, then it deferred its plan. The lock was held for the duration of
group commit, and was released once group commit completed. As a
consequence of this, only one write could ever be committed at once.

This commit fixes the lock contention so many writes can commit
together. Individual writes do not acquire the lock. Instead, the lock
is acquired by the group commit operation.

Fixes #13205