coord: Fix timestamp selection for read then write queries
Previously, read then write queries performed reads by explicitly
specifying an AS OF equal to the global read timestamp. There was no
validation that the selected timestamp was greater than or equal to the
since of the involved tables. This is usually not an issue, because the
global read timestamp is usually greater than or equal to since. This
is not true when logical compaction window is 0, in which case the
global read timestamp is less than since.

This commit modifies the timestamp selection for read then write
operations to be identical to timestamp selection for normal reads.
Normal timestamp selection already factors in the global timestamp for
user tables, so the timestamp selected should be the same for all
values of logical compaction window greater than 0.

Works towards fixing #13221