compute: remove Kafka sink
Kafka sinks will move to storage, compute will only sink dataflows to
storage collections in the future. This means we can remove any support
for Kafka from the computed code.

At the moment, when you create a Kafka sink in SQL, that still gets
passed to computed in the dataflow description, causing computed to
panic. We will change that in a follow-up.