expr: Add a mz_version_num() function
Modeled after Postgres' server_version_num , it returns the
environmentd version in the format XXYYYZZ where XX is the major
version, YYY is the minor verion and ZZ is the patch version.

Having the version returned as an interger allows scripts and
tests to use the value in comparisons and testdrive's $skip-if
directive without having to parse the human-readable form.