coord: register a source for persist sinks
This commit makes the coordinator announce a source to the storage
controller when creating persist sinks. This makes the storage
controller create a storage collection which we can sink into.

It will also later give us a way to read sinked data back through the
source. This doesn't work currently because we don't add a catalog item
for the source.