compute: make persist_sink compatible with persist_source
Storage's persist_source expects data in persist shards to be of type
`(SourceData, ())`, so compute's persist_sink needs to write data of
that type too. Prior to this commit, persist_sinks wrote data of type
`(Row, Row)`.

We perform the translation by dropping the key and wrapping the value
into `SourceData`. This is correct because persist_sinks (like tail
sinks) don't use keys, so only the sinked values are non-empty.

Kafka sinks are the only sinks that use keys. Once they get removed from
compute (and moved to storage), we will be able to get rid of the
awkward translation step.