dataflow-types: forward sink collection uppers
This commit makes the compute controller forward updates of sink storage
collection `upper`s to the storage controller. With this, the storage
controller is able to correctly advance the the collections' `since`
frontiers, unlocking compaction.

Eventually, we'll want the storage controller to keep track of storage
collection frontiers on its own, rendering this extra code in the
compute controller unnecessary.