persist: fix bug when ColumnarRecords is used with a very small max len
This only came up in tests where we set a very small max len, but it's
worth fixing. Before this fix, if the first updated added was larger
than the max len, we'd emit an empty ColumnarRecords. Now we don't.

Also, while we're in here, remove the Option wrapper from take_filled.
We don't get anything from wrapping it in an option.