persist: hook up an initial implementation of compaction
See the comments in trace.rs for an overview of how this is set up.

Compaction tests can quickly become a maintenance burden as it's
difficult to balance invariant verification against needing lots of
diffs when small compaction heuristics change. Invest early in
datadriven tests (the same style and underlying library as testdrive).

Add an `MZ_PERSIST_COMPACTION_ENABLED` env var as an escape hatch,
though as CI has shown, this is pretty inadvisable before we have
incremental state.