coord: Refactor RemovePendingPeeks command
All `Command`s correspond to some client initiated operation via one
of our APIs. The one exception to this is the `RemovePendingPeeks`
command, which is a message that the Coordinator sends to itself to
signal that some pending peek should be removed. This probably does not
belong in the `Command` enum and instead should be it's own message
type.