coord: Remove unnecessary background task (#13252)
This commit removes a background task that sends a message over a tokio
channel. Sending a message over a tokio channel is non-blocking and
doesn't need to be done on a background task. It's likely more work
to set up the task then it is to just send the message. It's likely
that this task used to do more, but it is no longer necessary as a
separate task.