storage: use Debezium sequence numbers for PostgreSQL deduplication
Since Debezium 1.5, the "sequence" and "lsn" fields of the PostgreSQL
source metadata has been sufficient to perform correct, constant-space
deduplication of PostgreSQL. This commit updates our PostgreSQL
deduplication code to rely exclusively on these fields.

The code previously relied on a the `transaction.total_order` field, but
that field was only available when users explicitly enabled the
transaction metadata topic.

The one wrinkle here is that updates to the primary key result in two
events at the same sequence number: a deletion of the old value and a
creation of the new value. This commit accounts for that by including
the operation type in the sequence number and sorting delete events
before create events.

Fix #7557.