storage: remove configurable debezium deduplication mode
We've long supported a "deduplication" parameter for `ENVELOPE DEBEZIUM`
that controls how Materialize attempts to detect and suppress duplicate
data in Debezium-enveloped Kafka topics. This parameter has been
undocumented for the entirety of its existence, and is syntactically in
the wrong spot to boot (#3823). To correctly set this parameter, users
needed to navigate a complicated and undocumented web of factors
involving the upstream database, the Debezium version, and the amount of
memory available to Materialize. In practice, users only set this option
if instructed to by a Materialize engineer.

This commit removes the option entirely, on the theory that
deduplication should happen automatically, use a reasonable amount of
memory, and produce correct results when used with a recent,
correctly-deployed Debezium connector.

Fix #3823 and fix #8691 by removing the feature.