coord: Remove background task from group commit
Previously, write transaction would start a background thread after
submitting a write for group commit. This background thread would wait
for group commit to finish and then send a response to the client.

This commit removes these background threads and modifies group commit
so that responses are sent directly in group commit. This slightly
cleans up some of the write transaction code.