test/testdrive: simplify some ancient tests
A number of our test scripts are from an era where Materialize required
Debezium-formatted data and explicit timestamp closing messages. This
commit simplifies those tests by using append-only Avro sources.

This reduces the amount of work required to update test scripts when we
change our Debezium handling--e.g., in #13263, when we depend on
additional fields that are provided by real Debezium but not by our test
scripts which approximate Debezium.