persist: remove per-reader sinces, add SinceHandle
The `SinceHandle` allows one holder to downgrade the since of a shard.
When creating a new `SinceHandle`, older `SinceHandles` are fenced off,
meaning they cannot downgrade the since anymore. This should be used by
the storage controller, and the fencing ensures that only one controller
can be active at a given time.