coord: Implement consistent timestamp recovery
The consistency of certain queries relies on the Coordinator to return
strictly increasing timestamps. The Coordinator returns timestamps with
the help of a global timestamp. Previously, when the Coordinator
started it initialized the global timestamp to the current system
clock. This is fine if the system clock always goes forward, but this
is not always the case. The system clock can go backwards or the
Coordinator can be restarted on a separate machine with a different
clock. As a consequence, it was possible for the global timestamp to
decrease after a restart.

This commit implements a timestamp recovery protocol that prevents the
global timestamp from decreasing after restarts. The global timestamp
is initialized to the max of all object's uppers.

Works toward resolving #11631