dataflow-types: move common client code to new mz-services crate
As described in #12211, it is time to split the dataflow-types crate
into the compute and storage halves. This commit is the next step
towards that goal. It moves the `GenericClient` family of traits and the
gRPC client/server traits into a new `mz-services` crate.

This commit is almost entirely pure code movement, with a few small
changes to documentation comments and log messages to avoid mentioning
too much about "storage" or "compute" in the general purpose
`mz-services` crate.