service: remove LocalClient and RemoteClient
These clients were roughly newtypes around Partitioned<ProcessLocal> and
Partitioned<GrpcClient>. Remove them in favor of using the inner types
directly. This avoids a lot of boilerplate, and should make it easier to
tackle #12946 because there's one less client in the stack.

The very minor downside is an extra box around the local clients in
computed and storaged, but it's hard to imagine this mattering.