service: rework reconnection logic
Rework how reconnection works in the client stack. Rather than a
complicated interplay of error returns and cancel safe reconnection
logic that was split between GrpcClient, Partitioned, and
ActiveReplication, make ActiveReplication the sole place where
reconnections are attempted. ActiveReplication is the only place where
failed connections can be properly handled, so this cleans up the logic
substantially, and in particular makes it possible to guarantee that all
communication with a compute cluster will actually begin with a
CreateInstance command.

It's still a complicated protocol between ActiveReplication and each
replica task, but at least the complexity is isolated to the single
module, and the `reconnect` logic in Partitioned and GrpcClient becomes
dead simple.

Fix #12233.
Fix #12645.
Fix #12843.
Fix #12993.

Touches #12946.