dataflow-types: split storage types into storage crate
This is the next step towards splitting the dataflow-types crate into
its storage and compute halves (#12211).

The split here isn't perfect. There are some types (e.g., `ReadPolicy`)
that we might like to live in some shared crate, and for now I've just
tossed them into the storage crate. We can do that fine tuning later,
once this massive refactor is out of the way.