dataflow-types: split into compute-client and controller
Split out the unified storage and compute controller into a new
`controller` crate, then rename the remainder of `dataflow-types` to
`compute-client`. I'm going to say that this fixes #12211. There are
still a few straggler types around, but hopefully it's now obvious both
when those types are misplaced and where they ought to live.