mz-coord: rename to mz-adapter
The mz-coord crate is the bulk of the platform adapter layer. Rename it,
to make the code align with our design documents.

The `Coordinator` remains as a type in the adapter layer, for now, to
avoid sullying the word "adapter" with the historical baggage of
`Coordinator`.

This commit is again pure code movement.