storage: add first pass implementation of reconciliation
Add a `StorageCommandReconciliation` struct that nearly exactly mirrors
the existing `ComputeCommandReconciliation` struct. This makes it
possible to restart `environmentd` without also restarting `storaged`.

There was a lingering bug in compute command reconciliation which would
cause frontier upper information to get out of sync during
reconciliation, as the `computed` process would not suppress frontier
upper responses before sending over the correction to account for the
new controller. Because storage reconciliation now shares the same code,
that bug presented as storage collections getting wedged at their
original timestamp after restart.

This commit fixes that bug by teaching the gRPC server to communicate
with the reconcilers when a new controller connects. The reconcilers can
then correctly suppress frontier upper responses until the controller
expresses interest in the frontier.

One glitch remains: there is no way in the protocol to communicate that
a source was dropped and compacted out of the command stream. We can
solve that in a follow up, though.

This complements the work in #13360; that work, it turns out, is about
rehydrating a `storaged` process when it crashes; this work is about
the opposite situation, when `environmentd` crashes but `storaged` stays
online.

Part of epic #12857.
Split out from #13360.