sql: switch all errors to PlanError
Most of the functions still produce only PlanError::Unstructured, but
this will allow us to incrementally move each error to a structured
error type.

This allows the error structure to be propagated upwards to the
coordinator when it exists, which will allow the structured SQL errors
to return hints and long form descriptions.

In particular, after this change, the error messages under construction
in #13445 will be able to properly emit hints.