Split out `ActiveReplication` state
We move everything but the replica task handles into their own
struct, of type `ActiveReplicationState`. This allows us to call
methods on that state even when the task handles are mutably borrowed.

The main purpose of this is to break large chunks of code out
into non-async methods, so that we may more easily check the code for
cancelation safety.