storage: track dependencies of storage collections/acquire read holds
Before, the storage controller didn't track dependencies that storage
collections might have on other storage collections. The only case where
we have this right now is Debezium sources with a transaction source
(which is gated behind the "unsafe" flag). In this case it is a bug,
because rendering the debezium source can fail when the transaction
source has been compacted (since has advanced).

The bug in Debezium sources did not surface because we currently don't
correctly downgrade the since of persist shards because we do not
properly expire all `ReadHandle`s. The bug will surface in multiple
failed tests on #13347, which removes per-reader sinces and therefore
makes us downgrade persist sinces correctly.

In the future, we will also need this for external sinks, which depend
on the storage collection that they are sinking from.

Closes #13490