persist: proto-ify state
This sets us up to be able to make backward compatible changes.

Remove some of the Serialize and Deserialize placeholder impls now that
we have a real serialization story. Leave the ones on SnapshotSplit
(which is intentional because we want it to impl timely::ExchangeData).
Also leave the ones on ShardId and PersistLocation because it seems like
these are being used outside persist (probably not on purpose) and so
it's better to save that for a follow up PR.