storage,compute: rethink reconciliation
Previously, reconciliation sat on *top* of storage/compute servers, and
attempted to paper over commands and responses so that clients believed
they were talking to a fresh server, while servers always believed they
were talking to the originally-connected client. This introduced a lot
of complicated logic and double bookkeeping.

This commit pushes reconciliation logic into the storaged and computed
servers. The servers are now aware of when a new client connects, and
they simply ditch any state that was related to the old client.

This tees us up for the last reconciliation task: garbage collecting
sources and dataflows that newly-connected clients are not interested
in.

Touches #10549.
Touches #12857.