sql: introduce and mandate PostgreSQL connections
Introduce a new PostgreSQL connection type, as in...

    CREATE CONNECTION pgconn FOR POSTGRES
      HOST postgres,
      DATABASE postgres,
      USER postgres,
      PASSWORD SECRET pgpass;

    CREATE SOURCE pg FOR POSTGRES CONNECTION pgconn;

The connection options correspond to the old libpq connection parameters
in the obvious way.

As decided in #13374, this commit also removes the old syntax for
specifying PostgreSQL connection details inline in the `CREATE SOURCE`
statement; connection details *must* now be specified in `CREATE
CONNECTION`.

Touches #11504.
Touches #13374.