service/grpc: mandate identical versions
Our initial version upgrade story will be to mandate that `storaged`,
`computed`, and `environmentd` are all running identical versions.
Upgrading `environmentd` already attempts to bring all `storaged` and
`computed` clusters up to the latest version, but it does so by relying
on Kubernetes to asynchronously restarting the `storaged` and `computed`
processes as the new version.

This commit adds a check in the gRPC communication layer for identical
versions. This will ensure that, during a version upgrade, the new
version of `environmentd` will not be able to communicate with and
confuse old `storaged` and `computed` processes until they are upgraded
to the correct version.

Fix #13040.