Initialize recording `since` frontier on creation
When creating a recorded view, initialize the target storage
collection's `since` frontier by advancing it to the recorded view's
`as_of`.

This ensures that readers are never allowed to read from the collection
at times between the minimum and `as_of`. This is desirable since we
know nothing about the state of the TVC at these times.