storage: for write frontier init for non-ingested collections
Non-ingested collections had their write frontier mutable contain
`(T::minimum(), 1)` but that was never retracted in the face of restarts
as the subsquent appends used the current shard upper. To fix this we
need to make sure the collection is initialized correctly so that
subseuquent `ChagneBatch` updates match up with the current state.

Fixes #13454

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>