storage: report initial write frontier
This commit extends the storage manager to take into account the initial
write frontier of (re-)created storage collections and report that
instead of the reported write frontier as long as the latter is still
behind.

This makes it possible for other components (like adapter) to get a
useful value for a collection's upper during bootstrapping, rather than
'0'.