coord: Use largest possible timestamp in dataflow
When we create a dataflow we specify an `as_of` that indicates from
which frontier onward must the results be valid. Ideally, we choose the
`as_of` as large as possible, but no larger. Not choosing the as_of as
large as possible has some negative performance implications.

Previously, we were not choosing an `as_of` as large as possible and
instead using the `since` of the associated IDs.

This commit modifies the `as_of` selection for dataflows to use the
largest readable timestamp.

Fixes #13615