coord: coord: Acquire read holds for all timelines (#13605)
This commit updates the Coordinator so that it will hold a read hold on
all objects at the current timestamp of the TimestampOracle for that
object. Periodically, the Coordinator will update all of these read
holds. This may have a negative effect on memory utilization of all
nodes since compaction is being held back. Also, this increases the
amount of communication between the Coordinator and all nodes which
might introduce additional latency.

The read holds help ensure that all objects on the same timeline are
joinable and they are needed in future work to help with strict
serializability.

Works toward resolving #12309