Store the introspection shards for replicas
This commit persists the persisted logs of replicas in the
stash/catalog.

It refactors also the persisted_logs, it is now contained in
ConcreteComputeInstanceReplicaConfig. The former
ConcreteComputeInstanceReplicaConfig is renamed to
ConcreteComputeInstanceReplicaLocation. A corresponding migration has
been implemented.

On an empty catalog, the persisted_logs of the default cluster
are set to Default. This value is then on first boot replaced with
an allocated default set of shards and written back to the catalog
storage.