compute: set a root tracing span for peek dataflows
Some dataflows live indefinitely (for example maintaining an INDEX), but
some are short lived (created on behalf of a peek) and we'd like
different tracing behavior for each. For the latter, we want to attach
all tracing to a single root span (so they can ultimately be connected
to the peek). But there shouldn't be a single root for long-lived
dataflows (our tracing library, and the tracing ecosystem in general, is
built around short-lived requests and responses).

Sniff between these two cases with a heuristic that assumes any dataflow
exporting a transient id is itself transient.

Combined with #12828 (or something like it) this would connect a single
trace all the way from a `SELECT` down to the persist calls fetching
data for it inside computed. Even before that, this commit is useful on
its own to group the persist activity for the peek dataflow under a
single root.