compute,storage: add optional dataflow execution spans
Create a root tracing span for the execution of each dataflow in storage
and compute. Each operator can, at its option, enter this span when it
is scheduled, and exit the span when it yields. This is intended for
operators which call into code that is instrumented with the `tracing`
crate, so that the spans created by that instrumented code get parented
to the span for dataflow that created them.

This commit has the same effect as #13634--see that pull request for a
screenshot of the resulting traces. I made the following tweaks to that
PR:

  * The execution span is explicitly plumbed through to the relevant
    operators, rather than being passed through thread-local state. I
    find this easier to reason about than thread-local state in
    timely-related code. For me, at least, the timely idiom of
    short-lived functions creating long-lived closures makes reasoning
    about TLS state quite difficult.

  * The execution span is created unconditionally. It seems harmless to
    create this span even for long-lived dataflows. If it *is* a
    problem, it's easy to add back the logic that only creates the span
    for transient dataflows.

  * The execution span "follows from" the render span, rather than being
    a direct child. I think this better aligns with the tracing model.
    Dataflow execution can be thought of as an asynchronous background
    task that is *triggered* by dataflow rendering, rather than work
    that is done *during* dataflow rendering.

  * The timely operator closure enters the span, rather than each future
    involved in the operator. This should be a bit more robust against
    future refactorings of the code losing calls to `.instrument`.

Alternative to #13634.