orchestrator-kubernetes: minimize match labels
Kubernetes does not permit StatefulSets to change their match labels
after creation. So restrict the match labels to the minimal set, to
reduce the frequency with which they change.

We'll likely need to get smart enough to delete and recreate
StatefulSets when necessary, but this commit should stave off the
problem for now. For example, I believe PR #13578 will not be an
additional breaking change on top of this PR.

When we deploy this change to cloud, we'll need to manually delete the
StatefulSets in all existing environments.

Touches MaterializeInc/cloud#3529.