sql: disable/remove FORMAT ... USING SCHEMA FILE
Cloud-native Materialize does not have access to files. So
remove/disable support for specifying Protobuf and Avro schemas in
files.

There are a few too many protobuf-based tests that use this right now,
so `FORMAT PROTOBUF USING SCHEMA FILE` remains as an unsafe mode feature
until we have time to update the tests.