coord: Add strict serializable isolation level (#13487)
Previously, Materialize only exposed a single isolation level of
serializable. Materialize could create a total order of all operations
using the timestamp of that operation. However, there was no guarantee
that a read was performed at an equal to or higher timestamp than all
previous reads.

This commit adds the strict serializable isolation level. To learn more
about isolation levels see: https://jepsen.io/consistency. All reads
within a single timeline is performed at a timestamp equal to or higher
than all previous reads.

We accomplish this by using a TimestampOracle to get timestamps for
all reads. The TimestampOracle is guaranteed to always provide
increasing timestamps. We know that the timestamp will be valid to use
for reads, because the Coordinator maintains read holds on these times.

This commit does not prevent strict serializability violations if there
are multiple Coordinators active at the same time. This is a known
issue and a fix is being designed/worked on.

Fixes #12309