expr: make `VisitChildren` generic
For mutually recursive data types (i.e. types consisting of N sorts
that mutually depend on each other, like `Relation` and `Scalar`
in `Hir`), it will be beneficial to be able to visit children of
different sorts.

To do this, I suggest to generalize `VisitChildren` with a type
parameter `T` indicating the type of children to be visited.

The blanket implementation of `Visit` still needs a `VisitChildren`
implementation for the same type (`T: VisitChildren<T>`).