coord: change read holds to use uppers
Previously read holds would use the global timestamp for each timeline
which is used when serving strictly serializable reads. We can allow non
strict serializable reads to always be immediately available as well
by using the least common upper instead of the global timestamp. This
is useful for expensive views that take a while to compute and would
otherwise always be behind the global timestamp.