expr: remove lifetime parameter from `VisitChildren`
Currently, the lifetime of the function parameters passed
to the various `VisitChildren` methods are constrained
to borrow for the duration of the enclosing receiver -
they are either `&'a mut Self` or `&'a Self`.

This differs from the `Visit` trait, which is (and indeed
it needs to be) more relaxed in this regard in order to
work.

This commit removes the lifetime bound, so `VisitChildren`
implementation for a certain child type `T` can rely on
`Visit` implementations for another type `U`.