coord: Split coord.rs into multiple modules
This commit attempts to split coord.rs into multiple sub-modules.
coord.rs has become extremely large and close to unmanageable. It is
hard for anyone to wrap their heads around everything that the
Coordinator tries to do. This refactor attempts to make interacting
with the Coordinator easier by separating it into logical parts. This
refactor is not perfect, but hopefully is a step in the right
direction.

An unfortunate side-effect of this is that the git history will become
very difficult to use for coord.rs.

This commit only does the following:
- Create new files.
- Rename files.
- Move methods/functions to a new file.
- Extract a block of logic into a function/method.
- Rename a function/method.
- Add doc comments.

Importantly, none of the actual logic has changed.