sqllogictest: add a --log-filter option
This allows disabling the spammy logs from the various computed and
storaged processes that are spawned during an SLT run.

Also rename the OpenTelemetryEnableCallback to TracingHandle to better
reflect its purpose. It's not actually a callback, but a handle that can
be used to manipulate the state of tracing. Accordingly, the `shutdown`
free function is moved onto the handle.