Permit hydration of an uninitialized cluster
A cluster that has never received a command before a client disconnects
is in a state where it has no old config. When a new client connects, the
hydration logic would assert that the new config would be equal to the old
one, but the test only makes sense of the cluster was initialized before.

Move the check into a condition to only assert if an old config exists. In
addition to this, add an assertion that if there is no previous config,
then there are also no old dataflows and frontiers.

This change fixes a previously unreported bug.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>