coord: Protect against multiple Coordinators
Previously, if multiple concurrent Coordinators were running at once,
there was a possibility to violate strict serializability. For example,
if one Coordinator wrote to a table at some timestamp and then another
Coordinator read from the table at an earlier timestamp.

This commit prevents strict serializability violations in the presence
of multiple concurrent Coordinators. Before sending a response to a
read transactions, the Coordinator will confirm that it's still the
leader. Only after confirming that it's the leader will a response be
sent.

There is a possibility that the Coordinator loses leadership after
checking but before sending a response. This is ok because any
transaction that the new Coordinator commits will be concurrent with
the read transaction of the previous Coordinator. Therefore, we are
free to order the concurrent transactions in any way.

New Coordinators will fence all tables from writes as a part of
bootstrapping by advancing them to the new Coordinator's timestamp.
Therefore, we do not need to check for leadership to commit a write.

Fixes #13461