compute,storage: abort communication tasks on drop
While a storage host or compute replica is connecting, it cannot listen
for channel close events that indicate that the storage host or compute
replica has been dropped.

This commit fixes that by holding onto a reference to the communication
task that is aborted when the storage host or compute replica is
dropped.

Fixes #13668.
Alternative to and therefore would close #13846.