coord: Remove compute instances from timelines (#13862)
Introspection source indexes are not treated as normal indexes and are
special cased everywhere in the code. When we delete a cluster, we
never actually drop the introspection source indexes. They just
disappear and we manually update the necessary system tables.

When a cluster is deleted we were never removing the introspection
source indexes from the metadata about timeline read holds. Therefore,
the Coordinator would try and update read holds on non-existing indexes
in non-existing clusters and would panic.

This commit cleans up the read hold state when a cluster is dropped.

Fixes #13857