coord: Keep log index metadata consistent
Previously, when a cluster was deleted we would also delete all of the
log indexes (A.K.A. introspection source indexes) contained in the
cluster from the catalog. However, when we deleted the introspection
source indexes from the catalog, we weren't removing it completely.
For example, it was never removed from the Schema objects.

This commit updates the logic of removing introspection source indexes
so that it's identical to removing any other item from the catalog.

Might fix #12760