storage: drop tick closure in source operator when finishing up
We were seeing deadlock when a `persist_source` operator took a
`MutexGuard` (storing it in the internal Future state machine state) and
then (because the stop token was dropped) the `tick` logic never got run
again. We might want to make this shutdown protocol more graceful, but
at the very least, start by dropping the `tick` closure so anything it's
holding on to gets dropped.

Closes #13888