Process orch.: gather system info if needed
Gathering system info takes around 200ms on my machine because it needs to
read the entirety of the /proc file system. However, we only need this info
if there might be a stale process around. Change the code to only populate
the system info if it needs it.

This change reduces the end-to-end latency of spawning child processes in
most cases by around 200ms. This is a low-impact change since this code
will only be exercised in development and test environments.

This change fixes an unreported problem.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>