persist: add hacky temporary impl of automatic read lease expiration
When a process holding a ReadHandle restarts without dropping it (the
common case currently in our deployments) this ReadHandle holds a
permanent capability on seqno (and presumably since). This prevents GC,
leaking data into blob and consensus like crazy.

To prevent this situation, this commit introduces a simple, hacky
version of the automatic read lease expiration while we wait for the
real one that Paul is writing soon.

Touches #13861 (hopefully fixes it)