persist: thread down a tokio runtime for use in blocking work
When the Tokio runtime is shut down, it can't abort the spawn_blocking
closure (because it doesn't have await points). But it turns off the
timer. So the call to tokio::time::sleep explodes. See
tokio-rs/tokio#4862 for a better explanation of this.

To work around this, use a suggestion from the Tokio docs: create a
separate Tokio runtime for CPU-intensive tasks.

Touches #13867
Touches #13930