persist: add a long-lived listener to txn-list-append maelstrom impl
Keep a long-lived listen, which is incrementally read as we go. Then
assert that it has the same data as the short-lived snapshot+listen in
`read`. This hopefully stresses slightly different parts of the system.

I was hoping this would catch the Listen bug that took a while to track
down when I was writing the compaction PR. Sadly, it doesn't seem to (I
reintroduced the bug locally), but I think it's worthwhile to merge
anyway.