persist: fix more blob leaks
GC wasn't correctly using BlobKey and PartialBlobKey because it operates
directly on the proto state. Blob delete has to be resilient to retries,
so it doesn't fail for missing keys. (However, this also means our
blob_delete metric is a lie, it currently only tracks how many times
it's called. This will be fixed in a followup.)

A second smaller leak happened when applying a compaction response to
the state machine resulted in a noop. Since the state doesn't reference
the result of compaction, nothing would know to delete it.

There seems to be at least one minor leak left even after these two
fixes, but it's still being investigated.

Regression test also left for a followup because it will so much easier
to write after #13920.