coord: Prioritize read latency over write latency
The goal of this commit is to make sure that the global timestamp never
advances past the upper of any table. Doing this will make sure that a
table read will never have to block and wait for an upper to advance,
which should improve the performance of many table reads. This
performance improvement comes at the cost of performance for table
writes and DDL.

This is accomplished by combining table advancements into group commit.
All writes will advance every other table to the same timestamp of the
write. Periodically empty group commits are triggered which include no
writes, but advances every table. The global timestamp is only advanced
after these table advancements to the timestamp of the advancement.