persist-client: add a per-writer idempotency token
Track the most recent compare-and-append seqno for each writer in the
persist state. This makes it straightforward for a writer to retry a
compare_and_append operation. When a compare-and-append operation fails
indeterminately, the writer loads the most recent state. If it sees the
idempotency token of the undetermined write in the state, it knows the
write definitely succeeded. If it sees a different idempotency token, it
knows the write definitely failed, either because the CAS network
operation never made it to the consensus backend, or because the CAS
raced and lost with a prior canceled write from the same writer.

Fix #12797.
Fix #13921.