stash: prepare and pipeline postgres statements
Prepare most queries so they are available for reuse. This stops
tokio_postgres from sending Parse and Close messages.

Pipeline some of the more performance-sensitive functions by dropping all
work into futures. The order of messages (I believe!) is still the same,
as that seems to be guaranteed at the time the future is created. This
allows tokio_postgres to send them all at once instead of waiting
responses for each. A large amount of our work can be done this way.