stash: prepare postgres statements
Prepare most queries so they are available for reuse. This stops
tokio_postgres from sending Parse and Close messages.

Benchmarked from the previous commit, this is slower. I do not have
an explanation for that beyond a guess that it is related to benchmark
iteration startup cost, and in production use will be faster due to not
sending Parse and Close messages.

```
Benchmarking transact
Benchmarking transact: Analyzing
transact                time:   [75.933 ms 77.923 ms 80.054 ms]
                        change: [+16.250% +23.448% +29.994%] (p = 0.00 < 0.05)
                        Performance has regressed.
```