stash: tweak postgres schema
Index on a field that is regularly SELECT'd, instead of one that is
DELETE'd. Gut feeling is that this is better. We could do both, but
adding a third index to write to seems expensive.

It is fine for existing stashes to keep the old index. Not a correctness
problem, and most of them will be destroyed at some soon time anyway.