persist: always log determinate error retries at debug
The determinate "could not serialize access" errors happen often enough
in dev (which uses Postgres) that it's impeding people's work. At the
same time, it's been a source of confusion for eng. The situation is
much better on CRDB and we have metrics coverage in prod, so this is
redundant enough that it's more hurtful than helpful. As a result,
intentionally ignore INFO_MIN_ATTEMPTS and always logs at debug.