persist: also update WriteHandle upper in fetch_recent_upper
This fixes a bug where calling fetch_recent_upper didn't also advance
the self.upper field that powers `WriteHandle::upper`.

Also added a note that we might not even need to cache upper on
WriteHandle, since it should be redundant with info that's already in
machine+state. Didn't make that change here because I wanted to get this
fix out without having to think through the implications of making a
larger change.