Add a new unexpected-type error to improve messages
In #13682, we spotted an incorrect error message -- the error said a
particular name did not exist, where in fact it did but just had an
unexpected type. Add a new error type to represent this, convert
a couple methods that ought to be using it, and remove the old type.