persist: fix bug where Listen wasn't always advancing since
Listen used the batch lower in a sort of hack to try to come up with
some frontier strictly less than the batch upper (so that it would be
safe to use in a downgrade_since call). This interacted poorly with a
performance optimization we have that squashes empty batches together.
The result is that in a shard that is getting no updates (just frontier
progress), the Listen would see a series of batches with the same lower
and increasing uppers. The idea is still seemingly sound (though remains
quite hacky), we just wanted to use the previous call to next's
frontier, not the batch lower.