coord: Advance timelines when creating new obj
If an object is created with an initial read frontier that is larger
than the current timestamp of the timeline that object belongs to, then
we must advance that timeline to the frontier of that object. This
commit updates the code to do that.

The alternative would be to wait until the timeline catches up to the
object and then add the object to the timeline.