tests: make test_tail_table_rw_timestamps resilient to FETCH ALL indeterminacy
The change that flattens batches into parts introduced the behavior (as
a side effect) that the source will initially downgrade to the `as_of`
even when the snapshot is empty. Before, there was no progress message
attached to a snapshot so we were only starting to downgrade once the
listen batches came in.

This early downgrade can change what gets returned from a cursor, so we
make the test more resilient to indeterminacy in fetching.