persist: fix "unexpected missing blob" bug
SUBTLE! If Machine::merge_res returns false, the blobs referenced in
compaction output are deleted so we don't leak them. Naively passing
back the value returned by State::apply_merge_res might give a false
negative in the presence of retries and Indeterminate errors.
Specifically, something like the following:

- We try to apply_merge_res, it matches.
- When apply_unbatched_cmd goes to commit the new state, the
  Consensus::compare_and_set returns an Indeterminate error (but
  actually succeeds). The committed State now contains references to
  the compaction output blobs.
- Machine::apply_unbatched_idempotent_cmd retries the Indeterminate
  error. For whatever reason, this time though it doesn't match
  (maybe the batches simply get grouped difference when deserialized
  from state, or more unavoidably perhaps another compaction
  happens).
- This now bubbles up applied=false to the caller, which uses it as a
  signal that the blobs in the compaction output should be deleted so
  that we don't leak them.
- We now contain references in committed State to blobs that don't
  exist.

The fix is to keep track of whether applied ever was true, even for a
compare_and_set that returned an Indeterminate error. This has the
chance of false positive (leaking a blob) but that's better than a
false negative (a blob we can never recover referenced by state). We
anyway need a mechanism to clean up leaked blobs because of process
crashes.

Closes #14043