[DRAFT] Change DataflowError type
Motivation: include enough information in errors to allow
them to be retracted from upsert sources.

The way we will eventually allow retraction (coming soon, in a
follow-up to this PR) will be `"bad key bytes", null` for an
undecodable key error, and `"good key", null` for an undecodable value
that was inserted for `"good key"`.

Thus, our error type needs the following properties:

1. It must contain the original bytes associated with a decoding error
2. It must distinguish between key decoding errors and value decoding
errors
3. It must clearly separate decoding errors from random other errors in
the source pipeline. (Because decoding errors are the only ones that
it makes meaningful sense to correct via retraction to an upsert source; other errors
either occur later in the pipeline or are non-definite.)

In the service of these goals, we make the following changes:

1. We add an `original_bytes` field to the `DecodeError` type, which
is populated whenever the concept of original bytes makes sense.
2. We split `DataflowError::DecodeError` into `KeyDecodeError` and
`ValueDecodeError`.
3. We move some random non-decoding-related things that had crept into `DecodeError`
over the years out into a new variant `EnvelopeError`.

CAVEAT (and why this is still draft): This changes Protobuf types in
non-backwards-compatible ways. I would very much like to land this
before we get real customers in Platform and ignore the compatibility
issues, rather than having to leave around a bunch of deprecated
fields in protobuf and have to add logic to deal with them on
conversion.

I think this is fine, because there is no real pedagogical value to
doing the proper migration, and we're still (barely!) still in the
window where it's okay to blow everything away, but I thought it might
be controversial so I'm giving people the chance to respond.