compute: remove linkerd workarounds from networking
This essentially reverts #12832, which I suspect is no longer necessary
now that we no longer use linkerd in Materialize Cloud. The easiest way
to find out is to try it while we're still in alpha. We can always
revert this PR if we rediscover the networking issue.

For good measure, this commit also removes the handshake magic, which is
not necessary now that timely communication is not multiplexed on the
same port as HTTP and SQL traffic.