Create introspection views over introspection sources for each replica (#14072)
Add views defined over introspection views with a replica postfix id. These views are automatically created with replica create and removed with replica drop. If there are dependendants, CASCADE has to be supplied.

The catalog storage type LoggingConfig is extended with a ConcreteViews
variant, which now tracks LogView to GlobalId mapping.