expr: add more general visit methods
The methods in the current `Visit` trait do not allow to share
mutable state between pre and post-visit callback (this is due
to the Rust mutable borrow constraints).

To overcome this limitation, introduce four new methods:

- `visit`,
- `visit_mut`,
- `try_visit`, and
- `try_visit_mut`.

Instead of callback lambdas, these methods take an instance
of a type `V` that implements a corresponding trait:

- `Visitor`,
- `VisitorMut`,
- `TryVisitor`, and
- `TryVisitorMut`.

Each of these traits consists of a pair of `pre_visit`
and `post_visit` methods that can share mutable context.

This is the first building block towards introducing a
unified framework for derivation of bottom-up attributes.

To make room for the `Visitor` name, the old (mod-private)
`Visitor` type has been renamed to `StackSafeVisit`. I have
some other remarks about namings, but I will open a separate
PR for that.