Clean up string interpolation in pyactivate script
This updates the script to use logging's built-in support for
interpolation. This is beneficial because the interpolation will only be
performed when the log level matches the log statement's level.

This additionally updates all uses of `str.format` to be f-strings,
except in the case where we encounter a Python distribution < 3.9, where
we'll want to support the old style.