Retractable upsert errors
Motivation: Fixes #14084. This will allow users to retract errored values from Upsert sources, which worked in legacy, but not in platform.

It will also allow users to retract errored _keys_, which as far as I'm aware has never been possible.

We introduce a new `UpsertError` type, and slot it into a variant of `DataflowError`. This type/variant represents the errors that the upsert error is able to recover from: either key decoding errors, or various type of errors in interpreting values (decoding, and evaluating predicates) which are associated with a particular key.

To support retracting key decode errors, we add a `raw: Vec<u8>` field, which represents the original value of the key. This serves two purposes: (1) It makes it so that key errors only compare as equal if they actually arose from the same bytes, making sure we don't accidentally retract one errored key with a retraction meant for a different errored key, and (2) gives us enough information to re-populate the upsert map on restart.

To support value decoding errors, we include the (correctly-decoded) associated key. This is also in service of repopulating the key/value map on restart.

I think we might be starting to violate some abstraction boundaries with this `DataflowError` type, and maybe somebody should have a think about how best to separate out definite from non-definite errors, and which ones fit at which layers of the pipeline... but the present code works now and solves a real problem, so I think it should be landed.