transform: handle more cases in `ThresholdElision`
Improve the matching capacity of `ThresholdElision` by
introducing the several changes.

First, non-negativity as now modeled as a `NonNegative`
struct rather than a function. `NonNegative` implements
`Visitor` in order to:
- call `NonNegative::derive` in order to derive
  attribute values in a bottom-up manner, and
- maintain an environment of derived outcomes for
  each `LocalId` in scope.

Second, create a new struct `ThresholdElisionAction`
which wraps a `NonNegative` instance and itself implements
`VisitorMut` in order to:
- delegate attribute derivation to the `NonNegative`
  instance (a form of chain of responsibility),
- call the transform `action` (now moved to this new
  type) after the `NonNegative` attribute has been
  derived.

This ensures that non-negative `LocalId` values that fall
outside of the scope of the `Threshold` that is matched at
the root of the transform `action` method are correctly
classified.

Also, because both attribute derivation and transformation
attempts work in a single pass, the implementation improves
the runtime behavior (sadly the asymptotic complexity is
still `O(n^2)` because of the `is_superset_of` call).