[sinks] Move sinks to operate in storaged (#14009)
As title: move sinks from running in computed to storaged. The framework mostly mirrors the implementation for sources where it can as they have some major parallels. Big changes include:

Splitting the major sink structs into Storage$ and Compute$ such as StorageSinkDesc and `ComputeSinkDesc
Remove the two-phase sink initialization. Previously, we used Tail (which is now only part of the ComputeSink) effectively as a dummy sink. We can revisit this later -- but, for now, creating a sink requires spinning up a new storaged which is slow and defeats the purpose of having some initialization happening off main thread -- and the validations we'd expect to hit form quite a small set
Sink metrics were secretly only for Kafka -- so they've just been removed from ComputeSinks
As a summary, the new sink operates by reading in from a persist collection and immediately feeding that into the kafka sink implementation
Importantly, this doesn't do everything we want for sinks in storaged. Because we're behind unsafe mode, I think it's okay to do in multiple pieces:

We'll need a follow-up to switch from using a sidecar kafka topic to embedding the information in the header. The sidecar is still doable here -- and doing both of these at once would be quite chaotic. [kafka sink] Move consistency info from sidecar topic to headers  #14203
Some Compute* types still live in storage -- we'll want to change that (should be purely code movement): Remove ComputeSink* structs etc from storaged package #14202
Get tables working again (this was very minor in the docs) -- and tables are currently disabled for reuse_topic sinks -- so this shouldn't be a big issue: Allow sinks to export tables #14200
Motivation
Big piece of #11503, moving sinks to storaged

Tips for reviewer
I think looking by individual commit could be helpful. It's limited because there's a pretty big boundary at where I did a split between ComputeSink and StorageSink in the naming -- it's hard for me to rebase things over that -- but I've tried to make commits since then self-contained