Thread through the worker count
This simply passes the storage worker count on to Kubernetes, so the pod
is started up with the right number of workers.

Note that this has nothing to do with the "scale" -- ie. the number of
pods that will be spun up. We're keeping that fixed at 1 for now for
simplicity.