storage: allow all storage collections to be written to a Sink
Notably, this now allows creating a Sink that has a Table as input.

This removes the check for transitive dependencies when `reuse_topic =
true` and also removes the (already dead before this) code that had to
do with transitive dependencies of sinks. This code was used before
Sinks where moved to run in a `storaged` and to work only off of storage
collections.

Closes #14200