storage: replace dropped ReadHandles in PersistWorker with placeholder
When someone drops a storage collection (aka downgrade_since to the
empty antichain) replace it's ReadHandle in PersistWorker with a
placeholder. This placeholder acts as no-op for any later
downgrade_since calls (which is required for idempotence in the face of
potential replays).

Before this commit, we were keeping the ReadHandle alive so that it
could no-op future downgrade_since calls, but this was causing a leak in
persist.

Touches #14251