sql: Fix overflow when negating min ints
This commit fixes a panic when trying to negate the minimum value of
some integer type. Integer types are not symmetric with respect to
their positive and negative ranges. There is one more negative number
than positive numbers. Therefore, negating the minimum value of an
integer type will fail.

Fixes #14291