sql: guard fallible cast expression
SQL doesn't guarantee the order of execution of expressions in the
SELECT part of query and the WHERE part of a query.

It can happen that the optimizer chooses to run the SELECT expressions
*before* applying the WHERE clause, even if that particular row would be
filtered out in the opposite case. This has the bad sideeffect of an
error being produced from a row that would have participated in the
output.

This PR changes the SELECT project of postgres CREATE VIEWS such that it
only evaluates the CAST expressions if the WHERE condition is true by
wrapping each CAST expression in a CASE statement.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>