persist: move SerdeLeasedBatch to persist-client::fetch
SerdeLeasedBatch is now an important component in our story around
exchanging LeasedBatches in a way that helps users prevent leaking
SeqNos. As such, its visibility needs to change, so as to not
require making it visible through the impl::encoding module.

This commit accomplishes that goal by moving it into the fetch
module, which is made public.