coord: Fix builtin schema migration create order
It is important for built in schema migrations to recreate all objects
in the new dependency order, that is parents before children.
Previously, we failed to do that for the following reasons.
- The visit order of BFS is not guaranteed to be sorted in topological
order.
- The dependencies between builtin objects may change after a restart.
The previous code was using the persisted, possibly inaccurate,
dependencies.

This commit fixes both of these issues so that objects are recreated in
the correct order.