Fix fast path index lookup for non-existent keys
If the key to look up doesn't exist, then `cursor.seek_key` in
`compute_state.rs` tries to seek to a non-existent key, and it
lands on a different key. Earlier, when we were not removing
literal constraints from filters, the filter after the lookup would
filter out such lookup results, but now we have to take care to not
produce such results in the lookup itself.