adapter: accept parameters in /api/sql endpoint
Accept user parameters in a PG Simple Query-friendly way.

This approach is inspired by PG, namely this sentence
from the description of the Extended Query protocol:

> Furthermore, additional features are available [in Extended Query],
> such as the possibility of supplying data values as separate
> parameters instead of having to insert them directly into a query
> string.

This design does a slightly more safe version of what PG suggests,
by using a visitor over the AST and inserting the cast expressions
wherever there are parameters.

For details on this design, see:
https://github.com/MaterializeInc/materialize/pull/14073