persist: flatten snapshot LeasedBatches into parts
LeasedBatch exists so that persist_source can try to evenly distribute
the work of fetching and decoding shard data. However, Batches are
unevenly sized and can be arbitrarily large. Instead, distribute _batch
parts_, which are smallish (currently <128MiB, hopefully more like 8-32
MiB after inc state lands) and more evenly sized.

This commit only switches Snapshot to distribute based on parts because
it was simple and will give an immediate win. We should follow up to
give the same treatment to Listen (and also rename LeasedBatch to
something like LeasedPart) in a followup, but it's tricker because of
the attached progress.