nix: add a flake.nix with a devshell that delegates to shell.nix
Handling of the flake.lock file is a bit tricky because nix wants to
force-add it to git when it does any changes to it. You can get around
that with sth like:

```shell
$ git update-index --assume-unchanged misc/nix/flake.lock
```

Use the devshell like this:

```shell
$ nix develop ./misc/nix -c <your shell>
```

Where I use fish for <your shell>. You can also leave out the -c
argument if the default shell is fine for you.

Bonus content:

You can prevent garbage collection of the devshell dependencies but
installing a profile (which acts as a GC root):

```shell
$ nix develop --profile path/to/some-profile ./misc/nix -c fish
```

For example, you could keep live profiles in a hidden directory in your
home.