compute,storage: disable spurious tonic recompiles
tonic-build v0.8 will emit Cargo rerun-if-changed directives that are
actively unhelpful: because we use `..` as the Protobuf include path, we
end up recompiling if any file in `src` changes.

This commit restores the behavior of tonic-build v0.7, where the build
script for a crate is rerun only if the files inside of that crate
change.

h/t to @lluki for noticing the issue and @antiguru for bisecting to the
offending commit.