repr: fork persist's ProtoU64Antichain for non-persist use
Some non-persist code was re-using ProtoU64Antichain, but this was meant
as an internal detail of persist and there's no particular advantage we
get out of reuse here. In fact, trying to share it was getting in the
way of Matt's upcoming PR to switch mz_repr::Timestamp to a newtype
wrapper around u64.

This is also an opportunity to switch ProtoU64Antichain to the new,
idiomatic RustType and ProtoType stuff (instead of the persist From impls
that predate them), so go ahead and do that while we're in here.