compute-client: move controller.rs to controller/mod.rs
There are unfortunately two ways to organize submodules in folders in
Rust:

  - root/
    | submodule.rs
    | submodule/child.rs
and
  - root/
    | submodule/mod.rs
    | submodule/child.rs

Inconsistencies arise when both of them are mixed.

In the compute crates with use the latter style almost exclusively, so
this commit resolves the one inconsistency that is controller.rs by
renaming it to controller/mod.rs.