adapter: Prevent OOM from expanding peek result (#14507)
This commit adds heuristics to try and prevent out of memory errors
that result from combining multiple peek responses from COMPUTE and
from expanding binary results into unary results. The commit adds a
configurable system variable for max result size in bytes. If a result
ever grows larger than that variable, then we return an error instead
 of continuing to try and expand the result.

The default value for max result size was chosen to be 1 GiB because
environmentd is configured to have 4GiB of RAM in cloud.

This does not prevent OOM when compute sends a single large message.

Note: Binary to unary expanding involves converting a vector of
(row, count) to a vector of rows.

Works towards resolving #13307
Works towards resolving #13371