adapter(api/sql): properly handle committing single statements
Previously, single statements were treated as explicit transactions,
which is necessary for e.g. DDL, which cannot run intransaction blocks.
However, when the reuqests exited with committing, they were rolled
back.

This commit instead performs a COMMIT at the end of every request,
unless its final response was an error, in which case it performs
a ROLLBACK.