.github: add an issue template for CI flakes
CI is flaky enough at the moment that filing an issue for each flake is
onerous. We can mitigate some of the pain with a new issue template,
which is primarily based around a Buildkite link to a failed build. This
template should take less than a minute to fill out, and hopefully
increases the number of these flakes that get filed.