storage: don't use async_op! macro in persist_source
Alternative to #14572 for discussion. All glory and honor to Aljoscha,
who wrote the original fix and also figured out the root cause. This PR
differs from that one in that it's a slightly more straightforward
(brainless) copy of the async_op logic into where it was called
(adjusting things as minimally as possible to still fix the bug).

When using async_op! the entire body is run anytime the operator is
invoked (because of new input, frontier changes, or because someone
feels like it). The body is costly because it does things like
registering a new read handle (!!), allocating buffers etc. With this
change, we only do that once, and then keep that one future around until
the operator is finished.

Before, the fetcher operator was scheduled a lot and was taking up 100%
of the CPU timer of its worker, even with no data coming in. After this
change, it mostly sits idle, as it should.

You can see the impact by looking at the introspection views, and by
observing that `computed` with a simple TAIL is not pegged at 100 %
usage anymore.

The below printouts are from running a TAIL on a table that doesn't see
new updates.

Without this fix:

```
materialize=> select mdo.id, mdo.name, mdo.worker, mse.elapsed_ns
from mz_scheduling_elapsed as mse,
     mz_dataflow_operators as mdo
where
    mse.id = mdo.id and
    mse.worker = mdo.worker
order by elapsed_ns desc;
 id  |                      name                       | worker | elapsed_ns
-----+-------------------------------------------------+--------+-------------
 328 | Dataflow: tail-t7                               |      0 | 29117620155
 341 | Dataflow: tail-t7                               |      0 | 28567992731
 330 | persist_source 0: part fetcher u1               |      0 | 27912365122
 329 | persist_source User(1): part distribution       |      0 |   298032358
 332 | persist_source User(1): consumed part collector |      0 |      817396
 338 | tail-t7                                         |      0 |      813745
 336 | InspectBatch                                    |      0 |      566229
 334 | OkErr                                           |      0 |      347089
(8 rows)

Time: 44.527 ms
```

With this fix:

```
materialize=> select mdo.id, mdo.name, mdo.worker, mse.elapsed_ns
from mz_scheduling_elapsed as mse,
     mz_dataflow_operators as mdo
where
    mse.id = mdo.id and
    mse.worker = mdo.worker
order by elapsed_ns desc;
  id  |                    name                    | worker | elapsed_ns
------+--------------------------------------------+--------+------------
 1078 | Dataflow: tail-t17                         |      0 |  111508724
 1091 | Dataflow: tail-t17                         |      0 |   96973716
 1079 | persist_source u1: part distribution       |      0 |   79733907
  266 | Dataflow: temp-view-t3                     |      0 |    6401428
  297 | Dataflow: temp-view-t5                     |      0 |    6299725
  295 | Dataflow: temp-view-t3                     |      0 |    5942598
  326 | Dataflow: temp-view-t5                     |      0 |    5841977
 1080 | persist_source u1: part fetcher 0/1        |      0 |    4820154
  268 | persist_source s263: part fetcher 0/1      |      0 |    2888588
  299 | persist_source s278: part fetcher 0/1      |      0 |    2782819
 1088 | tail-t17                                   |      0 |     671058
 1082 | persist_source u1: consumed part collector |      0 |     539166
 1086 | InspectBatch                               |      0 |     500139
 1084 | OkErr                                      |      0 |     308879
  285 | ArrangeBy[[Column(0), Column(1)]]          |      0 |      95100
  316 | ArrangeBy[[Column(0)]]                     |      0 |      80980
  290 | Map                                        |      0 |       2930
  321 | Map                                        |      0 |       2790
(18 rows)

Time: 39.292 ms
```