compute-client: introduce compute controller commands
This commit introduces the
`mz_compute_client::controller::durable::Command` enum that defines
valid manipulations of the durable compute controller state. It
re-defines the API of `DurableState` in terms of this new `Command`
type. Now there is only a single public method that takes `&mut self`
(`apply`) and is therefore allowed to mutate the durable state (modulo a
couple exceptions that need to either be removed or justified in
documentation). Subsequent commits will ensure that every command
passing through `apply` is durably recorded.